Refactor CreateQuizScreen: Replace Column with LazyColumn for improved performance and structure; extract navigation buttons into a new component
